remove automatic polyfilling
Polyfills should only be required by applications, not modules.  If
another module/app requires the polyfill babel will throw and break,
leaving the module unusable. Generally we would just use the runtime
transformer but that doesn't work for on the fly compilation. Still it
is better to add a note letting the user know they need to include any
polyfills for their code (a la React) then presume to pollute their
global scope. ALSO the polyfill is enormous, and probably overkill for
most people.